Move new runfiles library to `//go/runfiles`
Follows https://bazel.build/rules/deploying#runfiles_library.
